LayoutTests/css-typedom/sameobject-expected.txt

Tests that idl [SameObject] is specified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS element is still alive
PASS successfullyParsed is true

TEST COMPLETE
